requirements.txt
concrete-ml==1.5.0
concrete-python>=2.5.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.12.0